+++ plat/pc86/libsys/brk.c
@@ -28,7 +28,7 @@ int brk(void* newend)
 	return 0;
 }
 
-void* sbrk(intptr_t increment)
+void* sbrk(int increment)
 {
 	char* old;
